[antlir] Add mkfs.ext4 options to improve build speed
Summary:
Playing around with options that should theoretically improve the build creation time for CVM builds. This is on the critical path and I want to make sure that these options are applied, but also safely applied.

I'm going to make sure that CVM builds are triggered on the `ext4` packager, since we are the primary and only customer for this anyways.

Test Plan: Wait for CVM cogwheel tests to all go green. Then, in a few days, check the build graphs to see if we notice an improvement in build times.

Reviewed By: vmagro

Differential Revision: D86312712

fbshipit-source-id: 3583715f5dfd17191edb34bc0a080a5a6683e71e