[antlir2] Generate manifest and unprivileged dir in caf package
Summary:
This is necessary to support rootless CAF builds, as we must be able to determine the permissions for every entry in the subvol (some of which may require root), but we can't unshare during CAF builds (as at that point we're in a threaded context where unshare doesn't work, and it's very difficult to break that out).

As a result, we can simply generate the manifest as part of the CAF package action, and provide that generated manifest to CAF directly when uploading. This piggybacks on the relevant recent work of bunn to split out manifest generation in CAF.

Test Plan:
Added unit test, this will be further consumed in upcoming diffs in stack which will contain more e2e testing.

For this diff, changes should be purely additive as the default output of the package is unchanged, so I don't expect any functional differences.

Reviewed By: vmagro

Differential Revision: D86433415

fbshipit-source-id: 00b1475646c3fbaae00be7d1d9a53f880e30cec1