antlir: Fix failing distro toolchain tests
Summary: D87403958 broke these tests because `feature.install` now implicitly installs xarexec into the image if we're building with the distro toolchain. We need to (a) not install xarexec if we don't need it and (b) not implicitly install it in the image_python_test macro.

Test Plan:
CI

Tests pass:
https://www.internalfb.com/intern/testinfra/testrun/2533275121767042

Reviewed By: vmagro

Differential Revision: D87750611

fbshipit-source-id: 9440efa80356afdad8f29ebd102024049118e9fd