[antlir2][buck] create a wrapper for feature rules
Summary:
All feature rules must share certain attrs. Until this diff stack, that was
only the `plugin`, but now I want to add `location`, so I'm defining a macro
and converting all features to use it so that the rest of this stack is a bit
simpler.

Test Plan: waitforsandcastle

Reviewed By: naveedgol

Differential Revision: D87938675

fbshipit-source-id: b93a81f13276907a07923082c5a26146fd0f45fa